Move all EnergySystem unit tests to one file
